64位编译支持.
git-svn-id: https://kbengine.svn.sourceforge.net/svnroot/kbengine@413 d26bdfbe-e24b-4f15-971f-8e0596e1bc08

+++ kbe/src/lib/entitydef/entity_macro.hpp
@@ -77,7 +77,7 @@ public:																										\
 			return;																							\
 		}																									\
 																											\
-		int pos = 0;																						\
+		Py_ssize_t pos = 0;																					\
 		PyObject *key, *value;																				\
 		PyObject* pydict = PyObject_GetAttrString(this, "__dict__");										\
 		PyObject* cellDataDict = PyObject_GetAttrString(this, "cellData");									\
