+++ kbe/src/lib/pyscript/sequence.hpp
@@ -34,16 +34,16 @@ public:
 	Sequence(PyTypeObject* pyType, bool isInitialised = false);
 	virtual ~Sequence();
 	
-	static int seq_length(PyObject* self);
+	static Py_ssize_t seq_length(PyObject* self);
 	static PyObject* seq_concat(PyObject* self, PyObject* seq);
-	static PyObject* seq_repeat(PyObject* self, int n);
-	static PyObject* seq_item(PyObject* self, int index);
-	static PyObject* seq_slice(PyObject* self, int startIndex, int endIndex);
-	static int seq_ass_item(PyObject* self, int index, PyObject* value);
-	static int seq_ass_slice(PyObject* self, int index1, int index2, PyObject* oterSeq);
+	static PyObject* seq_repeat(PyObject* self, Py_ssize_t n);
+	static PyObject* seq_item(PyObject* self, Py_ssize_t index);
+	static PyObject* seq_slice(PyObject* self, Py_ssize_t startIndex, Py_ssize_t endIndex);
+	static int seq_ass_item(PyObject* self, Py_ssize_t index, PyObject* value);
+	static int seq_ass_slice(PyObject* self, Py_ssize_t index1, Py_ssize_t index2, PyObject* oterSeq);
 	static int seq_contains(PyObject* self, PyObject* value);
 	static PyObject* seq_inplace_concat(PyObject* self, PyObject* oterSeq);
-	static PyObject* seq_inplace_repeat(PyObject * self, int n);
+	static PyObject* seq_inplace_repeat(PyObject * self, Py_ssize_t n);
 	
 	int length(void){ return values_.size(); }
 	std::vector<PyObject*>& getValues(void){ return values_; }
