+++ kbe/src/lib/pyscript/vector2.cpp
@@ -198,14 +198,14 @@ PyObject* ScriptVector2::pyGetVectorLengthSquared()
 }
 
 //-------------------------------------------------------------------------------------
-int ScriptVector2::seq_length(PyObject* self)
+Py_ssize_t ScriptVector2::seq_length(PyObject* self)
 {
 	ScriptVector2* seq = static_cast<ScriptVector2*>(self);
 	return seq->length();
 }
 
 //-------------------------------------------------------------------------------------
-PyObject* ScriptVector2::seq_item(PyObject* self, int index)
+PyObject* ScriptVector2::seq_item(PyObject* self, Py_ssize_t index)
 {
 	if (index < 0 || VECTOR_SIZE <= index)
 	{
@@ -215,11 +215,11 @@ PyObject* ScriptVector2::seq_item(PyObject* self, int index)
 	}
 
 	ScriptVector2* sv = static_cast<ScriptVector2*>(self);
-	return PyFloat_FromDouble(sv->getVector()[index]);
+	return PyFloat_FromDouble(sv->getVector()[static_cast<int32>(static_cast<int>(index))]);
 }
 
 //-------------------------------------------------------------------------------------
-PyObject* ScriptVector2::seq_slice(PyObject* self, int startIndex, int endIndex)
+PyObject* ScriptVector2::seq_slice(PyObject* self, Py_ssize_t startIndex, Py_ssize_t endIndex)
 {
 	if(startIndex < 0)
 		startIndex = 0;
@@ -249,14 +249,14 @@ PyObject* ScriptVector2::seq_slice(PyObject* self, int startIndex, int endIndex)
 				break;
 			case 1:
 				pyResult = PyTuple_New(1);
-				PyTuple_SET_ITEM(pyResult, 0, PyFloat_FromDouble(sv->getVector()[startIndex]));
+				PyTuple_SET_ITEM(pyResult, 0, PyFloat_FromDouble(sv->getVector()[static_cast<int>(startIndex)]));
 				break;
 			case 2:
 			{
 				Vector2 v;
 				
 				for(int i = startIndex; i < endIndex; i++){
-					v[i - startIndex] = my_v[i];
+					v[i - static_cast<int>(startIndex)] = my_v[i];
 				}
 
 				pyResult = new ScriptVector2(v);
@@ -266,7 +266,7 @@ PyObject* ScriptVector2::seq_slice(PyObject* self, int startIndex, int endIndex)
 			{
 				Vector3 v;
 				for (int i = startIndex; i < endIndex; i++){
-					v[i - startIndex] = my_v[i];
+					v[i - static_cast<int>(startIndex)] = my_v[i];
 				}
 
 				pyResult = new ScriptVector3(v);
@@ -282,7 +282,7 @@ PyObject* ScriptVector2::seq_slice(PyObject* self, int startIndex, int endIndex)
 }
 
 //-------------------------------------------------------------------------------------
-int ScriptVector2::seq_ass_item(PyObject* self, int index, PyObject* value)
+int ScriptVector2::seq_ass_item(PyObject* self, Py_ssize_t index, PyObject* value)
 {
 	ScriptVector2* sv = static_cast<ScriptVector2*>(self);
 
@@ -294,7 +294,7 @@ int ScriptVector2::seq_ass_item(PyObject* self, int index, PyObject* value)
 	}
 
 	Vector2& v = sv->getVector();
-	v[index] = float(PyFloat_AsDouble(value));
+	v[static_cast<int>(index)] = float(PyFloat_AsDouble(value));
 	return 0;
 }
