+++ kbe/src/lib/pyscript/vector3.cpp
@@ -201,14 +201,14 @@ PyObject* ScriptVector3::pyGetVectorLengthSquared()
 }
 
 //-------------------------------------------------------------------------------------
-int ScriptVector3::seq_length(PyObject* self)
+Py_ssize_t ScriptVector3::seq_length(PyObject* self)
 {
 	ScriptVector3* seq = static_cast<ScriptVector3*>(self);
 	return seq->length();
 }
 
 //-------------------------------------------------------------------------------------
-PyObject* ScriptVector3::seq_item(PyObject* self, int index)
+PyObject* ScriptVector3::seq_item(PyObject* self, Py_ssize_t index)
 {
 	if (index < 0 || VECTOR_SIZE <= index)
 	{
@@ -218,11 +218,11 @@ PyObject* ScriptVector3::seq_item(PyObject* self, int index)
 	}
 
 	ScriptVector3* sv = static_cast<ScriptVector3*>(self);
-	return PyFloat_FromDouble(sv->getVector()[index]);
+	return PyFloat_FromDouble(sv->getVector()[static_cast<int>(index)]);
 }
 
 //-------------------------------------------------------------------------------------
-PyObject* ScriptVector3::seq_slice(PyObject* self, int startIndex, int endIndex)
+PyObject* ScriptVector3::seq_slice(PyObject* self, Py_ssize_t startIndex, Py_ssize_t endIndex)
 {
 	if(startIndex < 0)
 		startIndex = 0;
@@ -252,14 +252,14 @@ PyObject* ScriptVector3::seq_slice(PyObject* self, int startIndex, int endIndex)
 				break;
 			case 1:
 				pyResult = PyTuple_New(1);
-				PyTuple_SET_ITEM(pyResult, 0, PyFloat_FromDouble(sv->getVector()[startIndex]));
+				PyTuple_SET_ITEM(pyResult, 0, PyFloat_FromDouble(sv->getVector()[static_cast<int>(startIndex)]));
 				break;
 			case 2:
 			{
 				Vector2 v;
 				
 				for(int i = startIndex; i < endIndex; i++){
-					v[i - startIndex] = my_v[i];
+					v[i - static_cast<int>(startIndex)] = my_v[i];
 				}
 
 				pyResult = new ScriptVector2(v);
@@ -269,7 +269,7 @@ PyObject* ScriptVector3::seq_slice(PyObject* self, int startIndex, int endIndex)
 			{
 				Vector3 v;
 				for (int i = startIndex; i < endIndex; i++){
-					v[i - startIndex] = my_v[i];
+					v[i - static_cast<int>(startIndex)] = my_v[i];
 				}
 
 				pyResult = new ScriptVector3(v);
@@ -285,7 +285,7 @@ PyObject* ScriptVector3::seq_slice(PyObject* self, int startIndex, int endIndex)
 }
 
 //-------------------------------------------------------------------------------------
-int ScriptVector3::seq_ass_item(PyObject* self, int index, PyObject* value)
+int ScriptVector3::seq_ass_item(PyObject* self, Py_ssize_t index, PyObject* value)
 {
 	ScriptVector3* sv = static_cast<ScriptVector3*>(self);
 
@@ -297,7 +297,7 @@ int ScriptVector3::seq_ass_item(PyObject* self, int index, PyObject* value)
 	}
 
 	Vector3& v = sv->getVector();
-	v[index] = float(PyFloat_AsDouble(value));
+	v[static_cast<int>(index)] = float(PyFloat_AsDouble(value));
 	return 0;
 }
